+++ src/main/java/com/foodie/portal/order/command/PayOrderCommand.java
@@ -8,5 +8,5 @@ import java.math.BigDecimal;
 @Data
 public class PayOrderCommand {
     @NotNull(message = "支付金额不能为空")
-    private BigDecimal paidPrice;
+    private double paidPrice;
 }
